# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for eCRF Application",
        "title": "eCRF API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/api/center/add": {
            "post": {
                "description": "add new center",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Center"
                ],
                "summary": "Add new  center",
                "parameters": [
                    {
                        "description": "addcenter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.AddCenter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/center/all": {
            "get": {
                "description": "getting all  centers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Center"
                ],
                "summary": "Get all centers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Center"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/center/delete": {
            "delete": {
                "description": "delete center",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Center"
                ],
                "summary": "Delete center",
                "parameters": [
                    {
                        "description": "deletecenter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteCenter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/center/name/{id}": {
            "get": {
                "description": "getting center name by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Center"
                ],
                "summary": "Get center name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/center/update": {
            "patch": {
                "description": "update center",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Center"
                ],
                "summary": "Update center",
                "parameters": [
                    {
                        "description": "updatecenter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateCenter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/journal/all": {
            "get": {
                "description": "gettning jouranl of operations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journal"
                ],
                "summary": "Journal Get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.GetJournal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
        "/api/subject/action/updatecolor": {
            "patch": {
                "description": "UpdateColor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject.Action"
                ],
                "summary": "UpdateColor",
                "parameters": [
                    {
                        "description": "UpdateColorRequest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AddProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/subject/action/updatecolorwithcomment": {
            "patch": {
                "description": "UpdateColorWithComment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject.Action"
                ],
                "summary": "UpdateColorWithComment",
                "parameters": [
                    {
                        "description": "UpdateColorWithCommentRequest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateColorWithCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AddProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/subject/action/updatefield": {
            "patch": {
                "description": "UpdateFieldValue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subject.Action"
                ],
                "summary": "UpdateFieldValue",
                "parameters": [
                    {
                        "description": "UpdateFieldValue",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateValueWithColor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AddProtocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requests.AddCenter": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "requests.DeleteCenter": {
            "type": "object",
            "properties": {
                "center_id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.UpdateCenter": {
            "type": "object",
            "properties": {
                "center_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "requests.UpdateColorRequest": {
            "type": "object",
            "properties": {
                "field_name": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                },
                "subject_number": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "requests.UpdateColorWithCommentRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "sendersrole": {
                    "type": "string"
                },
                "subject_number": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "requests.UpdateValueWithColor": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "integer"
                },
                "field_name": {
                    "type": "string"
                },
                "protocol_id": {
                    "type": "integer"
                },
                "subject_number": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "responses.AddProtocol": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Center": {
            "type": "object",
            "properties": {
                "center_id": {
                    "type": "integer"
                },
                "center_name": {
                    "type": "string"
                }
            }
        },
        "responses.DeleteProtocol": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.GetJournal": {
            "type": "object",
            "properties": {
                "actionType": {
                    "type": "string"
                },
                "body": {},
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "initiatorRole": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Cookie",
            "in": "Cookie"
        }
    }
}